requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
language-tool-python==2.7.1
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
python-pptx==0.6.21
requests==2.27.1
tqdm==4.64.0
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
XlsxWriter==3.0.3
zipp==3.8.0
